Added support in the ModelLoader for the remaining aspects of the file format, e.g. friction, gravity, connections and key connections, so the ModelLoader should now be (more or less) complete.
I've implemented the functionality in GBallModel that adds connections and key connections, and sets gravity and friction. I've starte to implement the method that checks if it safe to add an iGizmo in to the model (i.e. checking the coordinates aren't the same), so this still has to be finished along with the implementations for the addGizmo, addAbsorber and addBall. Once these have been implemented, the model for prototype 4 should be (more or less) complete.